requirements.txt
streamlit
pytesseract
pillow
deep-translator
opencv-python
gTTS